+++ src/dialogs/voice-assistant-setup/voice-assistant-setup-step-cloud.ts
@@ -4,6 +4,7 @@ import type { HomeAssistant } from "../../types";
 import "./cloud/cloud-step-intro";
 import "./cloud/cloud-step-signin";
 import "./cloud/cloud-step-signup";
+import type { HASSDomEvent } from "../../common/dom/fire_event";
 import { fireEvent } from "../../common/dom/fire_event";
 import { STEP } from "./voice-assistant-setup-dialog";
 
@@ -12,6 +13,8 @@ export class HaVoiceAssistantSetupStepCloud extends LitElement {
   @property({ attribute: false }) public hass!: HomeAssistant;
 
   @state() private _state: "SIGNUP" | "SIGNIN" | "INTRO" = "INTRO";
+
+  @state() private _email?: string;
 
   protected override render() {
     if (this._state === "SIGNUP") {
@@ -23,6 +26,7 @@ export class HaVoiceAssistantSetupStepCloud extends LitElement {
     if (this._state === "SIGNIN") {
       return html`<cloud-step-signin
         .hass=${this.hass}
+        .email=${this._email}
         @cloud-step=${this._cloudStep}
       ></cloud-step-signin>`;
     }
@@ -32,13 +36,17 @@ export class HaVoiceAssistantSetupStepCloud extends LitElement {
     ></cloud-step-intro>`;
   }
 
-  private _cloudStep(ev) {
+  private _cloudStep(ev: HASSDomEvent<HASSDomEvents["cloud-step"]>) {
     if (ev.detail.step === "DONE") {
       fireEvent(this, "next-step", {
         step: STEP.PIPELINE,
         noPrevious: true,
       });
       return;
+    }
+    // Carried between the steps so switching does not lose a typed address.
+    if (ev.detail.email !== undefined) {
+      this._email = ev.detail.email;
     }
     this._state = ev.detail.step;
   }
@@ -49,6 +57,6 @@ declare global {
     "ha-voice-assistant-setup-step-cloud": HaVoiceAssistantSetupStepCloud;
   }
   interface HASSDomEvents {
-    "cloud-step": { step: "SIGNUP" | "SIGNIN" | "DONE" };
+    "cloud-step": { step: "SIGNUP" | "SIGNIN" | "DONE"; email?: string };
   }
 }
