+++ src/panels/config/cloud/login/cloud-login-panel.ts
@@ -1,6 +1,7 @@
 import type { TemplateResult } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, query } from "lit/decorators";
+import type { HASSDomEvent } from "../../../../common/dom/fire_event";
 import { fireEvent } from "../../../../common/dom/fire_event";
 import { navigate } from "../../../../common/navigate";
 import "../../../../components/ha-alert";
@@ -71,11 +72,11 @@ export class CloudLoginPanel extends LitElement {
     cloudLogin.emailField?.focus();
   }
 
-  private _handleForgotPassword() {
+  private _handleForgotPassword(
+    ev: HASSDomEvent<HASSDomEvents["cloud-forgot-password"]>
+  ) {
     this._dismissFlash();
-    fireEvent(this, "cloud-email-changed", {
-      value: this._cloudLoginElement?.emailField?.value ?? this.email ?? "",
-    });
+    fireEvent(this, "cloud-email-changed", { value: ev.detail.email });
     navigate("/config/cloud/forgot-password");
   }
